requirements.txt
openai
python-dotenv
PyMuPDF 
jsonschema 
pydantic
